fix malfunctioned pid - previously i term not functioning due to mishandled relation between pos-pid and vel-pid
Previously I failed to figure out clearly when to clean up the
accumulated errors of i & d term of vel-pid when the velocity is set
by pos-pid. So I simply disabled the i and d in vel-pid and only enable
them when the setpoint of vel-pid is set by user directly. But I also
set a max vel output limit for pos-pid. so even pos-pid has i & d
enabled, they can't do much since there is a cap of their output